# public/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"IInstitution": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"institutionType": {
						"type": "string"
					},
					"level": {
						"type": "string"
					},
					"hasPharmacy": {
						"type": "boolean"
					},
					"admin": {
						"properties": {
							"email": {
								"type": "string"
							},
							"phone": {
								"type": "string"
							},
							"name": {
								"type": "string"
							}
						},
						"required": [
							"email",
							"phone",
							"name"
						],
						"type": "object"
					},
					"details": {
						"properties": {
							"percentage": {
								"type": "number",
								"format": "double"
							},
							"TIN": {
								"type": "string"
							},
							"location": {
								"type": "string"
							}
						},
						"required": [
							"percentage",
							"TIN",
							"location"
						],
						"type": "object"
					},
					"institutionId": {
						"type": "string",
						"nullable": true
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"name",
					"institutionType",
					"level",
					"hasPharmacy",
					"admin",
					"details",
					"createdAt"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"IServiceAct": {
				"properties": {
					"id": {
						"type": "string"
					},
					"label": {
						"type": "string"
					},
					"desc": {
						"type": "string"
					},
					"price": {
						"type": "number",
						"format": "double"
					},
					"serviceId": {
						"type": "string"
					},
					"institutionId": {
						"type": "string"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"service": {
						"$ref": "#/components/schemas/IService"
					},
					"institution": {
						"$ref": "#/components/schemas/IInstitution"
					}
				},
				"required": [
					"id",
					"label",
					"desc",
					"price",
					"serviceId",
					"institutionId",
					"createdAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IService": {
				"properties": {
					"id": {
						"type": "string"
					},
					"label": {
						"type": "string"
					},
					"desc": {
						"type": "string"
					},
					"level": {
						"type": "string"
					},
					"assignDuringOrientation": {
						"type": "boolean"
					},
					"institutionId": {
						"type": "string"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"acts": {
						"items": {
							"$ref": "#/components/schemas/IServiceAct"
						},
						"type": "array"
					},
					"institution": {
						"$ref": "#/components/schemas/IInstitution"
					}
				},
				"required": [
					"id",
					"label",
					"desc",
					"level",
					"assignDuringOrientation",
					"institutionId",
					"createdAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IServiceActResponse": {
				"properties": {
					"rows": {
						"items": {
							"$ref": "#/components/schemas/IServiceAct"
						},
						"type": "array"
					}
				},
				"required": [
					"rows"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IPaged_IServiceActResponse_": {
				"properties": {
					"data": {
						"$ref": "#/components/schemas/IServiceActResponse"
					},
					"totalItems": {
						"type": "number",
						"format": "double"
					},
					"currentPage": {
						"type": "number",
						"format": "double"
					},
					"itemsPerPage": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"data",
					"totalItems",
					"currentPage",
					"itemsPerPage"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"IServiceActRequest": {
				"properties": {
					"label": {
						"type": "string"
					},
					"institutionId": {
						"type": "string"
					},
					"desc": {
						"type": "string"
					},
					"price": {
						"type": "number",
						"format": "double"
					},
					"serviceId": {
						"type": "string"
					},
					"service": {
						"$ref": "#/components/schemas/IService"
					},
					"institution": {
						"$ref": "#/components/schemas/IInstitution"
					},
					"id": {
						"type": "string"
					}
				},
				"required": [
					"label",
					"institutionId",
					"desc",
					"price",
					"serviceId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IPriceChange": {
				"properties": {
					"price": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"price"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IServiceResponse": {
				"properties": {
					"rows": {
						"items": {
							"$ref": "#/components/schemas/IService"
						},
						"type": "array"
					}
				},
				"required": [
					"rows"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IPaged_IServiceResponse_": {
				"properties": {
					"data": {
						"$ref": "#/components/schemas/IServiceResponse"
					},
					"totalItems": {
						"type": "number",
						"format": "double"
					},
					"currentPage": {
						"type": "number",
						"format": "double"
					},
					"itemsPerPage": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"data",
					"totalItems",
					"currentPage",
					"itemsPerPage"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"IServiceRequest": {
				"properties": {
					"label": {
						"type": "string"
					},
					"institutionId": {
						"type": "string"
					},
					"desc": {
						"type": "string"
					},
					"institution": {
						"$ref": "#/components/schemas/IInstitution"
					},
					"level": {
						"type": "string"
					},
					"assignDuringOrientation": {
						"type": "boolean"
					},
					"acts": {
						"items": {
							"$ref": "#/components/schemas/IServiceAct"
						},
						"type": "array"
					},
					"id": {
						"type": "string"
					}
				},
				"required": [
					"label",
					"institutionId",
					"desc",
					"level",
					"assignDuringOrientation"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"IInstitutionExam": {
				"properties": {
					"id": {
						"type": "string"
					},
					"institutionId": {
						"type": "string"
					},
					"examId": {
						"type": "string"
					},
					"price": {
						"type": "number",
						"format": "double"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"institutionId",
					"examId",
					"price",
					"createdAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IInsuranceExam": {
				"properties": {
					"id": {
						"type": "string"
					},
					"institutionId": {
						"type": "string"
					},
					"examId": {
						"type": "string"
					},
					"price": {
						"type": "number",
						"format": "double"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"institutionId",
					"examId",
					"price",
					"createdAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IExam": {
				"properties": {
					"id": {
						"type": "string"
					},
					"exam_code": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"price": {
						"type": "number",
						"format": "double"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"institutionExam": {
						"items": {
							"$ref": "#/components/schemas/IInstitutionExam"
						},
						"type": "array"
					},
					"insuranceExam": {
						"items": {
							"$ref": "#/components/schemas/IInsuranceExam"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"exam_code",
					"description",
					"name",
					"price",
					"createdAt"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"IDrugCategory": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"name",
					"createdAt"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"jwtAuth": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT",
				"description": "Bearer token to access these API endpoints"
			}
		}
	},
	"info": {
		"title": "digital-med",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {
			"name": "Reponse Ashimwe"
		}
	},
	"paths": {
...
		"/api/service-acts": {
			"get": {
				"operationId": "GetAll",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IPaged_IServiceActResponse_"
								}
							}
						}
					}
				},
				"tags": [
					"Service Acts"
				],
				"security": [
					{
						"jwtAuth": []
					}
				],
				"parameters": []
			},
			"post": {
				"operationId": "Create",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IServiceAct"
								}
							}
						}
					}
				},
				"tags": [
					"Service Acts"
				],
				"security": [
					{
						"jwtAuth": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IServiceActRequest"
							}
						}
					}
				}
			}
		},
		"/api/service-acts/{id}": {
			"get": {
				"operationId": "GetOne",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IServiceAct"
								}
							}
						}
					}
				},
				"tags": [
					"Service Acts"
				],
				"security": [
					{
						"jwtAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "Update",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "boolean"
								}
							}
						}
					}
				},
				"tags": [
					"Service Acts"
				],
				"security": [
					{
						"jwtAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IServiceActRequest"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "Delete",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "boolean"
								}
							}
						}
					}
				},
				"tags": [
					"Service Acts"
				],
				"security": [
					{
						"jwtAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/service-acts/all": {
			"get": {
				"operationId": "GetNPaged",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/IServiceAct"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Service Acts"
				],
				"security": [
					{
						"jwtAuth": []
					}
				],
				"parameters": []
			}
		},
		"/api/service-acts/{id}/prices": {
			"put": {
				"operationId": "UpdatePrice",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "boolean"
								}
							}
						}
					}
				},
				"tags": [
					"Service Acts"
				],
				"security": [
					{
						"jwtAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IPriceChange"
							}
						}
					}
				}
			}
		},
		"/api/services": {
			"get": {
				"operationId": "GetAll",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IPaged_IServiceResponse_"
								}
							}
						}
					}
				},
				"tags": [
					"Services"
				],
				"security": [
					{
						"jwtAuth": []
					}
				],
				"parameters": []
			},
			"post": {
				"operationId": "Create",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IService"
								}
							}
						}
					}
				},
				"tags": [
					"Services"
				],
				"security": [
					{
						"jwtAuth": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IServiceRequest"
							}
						}
					}
				}
			}
		},
		"/api/services/{id}": {
			"get": {
				"operationId": "GetOne",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IService"
								}
							}
						}
					}
				},
				"tags": [
					"Services"
				],
				"security": [
					{
						"jwtAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "Update",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "boolean"
								}
							}
						}
					}
				},
				"tags": [
					"Services"
				],
				"security": [
					{
						"jwtAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IServiceRequest"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "Delete",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "boolean"
								}
							}
						}
					}
				},
				"tags": [
					"Services"
				],
				"security": [
					{
						"jwtAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/services/all": {
			"get": {
				"operationId": "GetNPaged",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/IService"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Services"
				],
				"security": [
					{
						"jwtAuth": []
					}
				],
				"parameters": []
			}
		},
...
		"/api/exams/all": {
			"get": {
				"operationId": "GetNPaged",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/IExam"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Exams"
				],
				"security": [
					{
						"jwtAuth": []
					}
				],
				"parameters": []
			}
		},
		"/api/exams/{id}/prices": {
			"put": {
				"operationId": "UpdatePrice",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "boolean"
								}
							}
						}
					}
				},
				"tags": [
					"Exams"
				],
				"security": [
					{
						"jwtAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IPriceChange"
							}
						}
					}
				}
			}
		},
...
		"/api/drug-categories/all": {
			"get": {
				"operationId": "GetNPaged",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/IDrugCategory"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Drugs category"
				],
				"security": [
					{
						"jwtAuth": []
					}
				],
				"parameters": []
			}
		},
...
		"/api/drugs/{id}/prices": {
			"put": {
				"operationId": "UpdatePrice",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "boolean"
								}
							}
						}
					}
				},
				"tags": [
					"Drugs"
				],
				"security": [
					{
						"jwtAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IPriceChange"
							}
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}